Rework evaluation and SValue type variant.
SValue variant no longer has list of children.

The list of children is encoded in the variant. Same for Qexpr.

This allows values such as int not have dummy children list.

Added support for Lambda variadics.

Evaluation of SValues now reduces the S-expression.

For example, consume the operator, pass the rest of the arguments to a
function. That function does not need to worry about an operator.